delete and create productCard page
